+++ src/elisa_client_api/elisa.py
@@ -19,10 +19,8 @@
 # 11/Feb/2013: add option to show attributes when searching for messages.
 #--------------------------------------------------------------------------------------
 
-from __future__ import absolute_import
-from builtins import object
-from elisa_client_api.core.restServer import RestServer
-from elisa_client_api.core.authentication import Authentication
+from .restServer import RestServer
+from .authentication import Authentication
 
 
 class Elisa(object):
