+++ src/elisa_client_api/request.py
@@ -18,17 +18,12 @@
 # 08/Jan/2012: add authentication
 #--------------------------------------------------------------------------------------
 
-from __future__ import absolute_import
-from future import standard_library
-standard_library.install_aliases()
-from builtins import str
-from builtins import object
 import urllib.request, urllib.parse, urllib.error
 import requests
 import os
 import mimetypes
 
-from elisa_client_api.exception import RestServerError
+from .exception import RestServerError
 
 
 class Request(object):
