+++ src/elisa_client_api/searchCriteria.py
@@ -19,10 +19,7 @@
 # 04/Dec/2012: use properties.
 #--------------------------------------------------------------------------------------
 
-from __future__ import absolute_import
-from builtins import str
-from builtins import object
-from elisa_client_api.core.searchField import SearchField
+from core.searchField import SearchField
 
 
 class SearchCriteria(object):
